UPSTREAM:DP wifi:ath12k: Set BIT(9) in pre-reset start command based on reset context
As part of the UMAC reset sequence, the host sends a pre-reset start
command to the firmware. BIT(9) in this command is used to indicate
whether the reset is due to a UMAC hang.

This change ensures that BIT(9) is set appropriately based on the type
of reset:
- For SSR (Subsystem Restart) recovery, BIT(9) is cleared.
- For other UMAC reset scenarios (e.g., Dynamic WSI Bypass), BIT(9) is set.

The decision is made based on the UMAC reset RX action received from
the firmware.

+++ backports-6.1-9a0dddfb3/drivers/net/wireless/ath/ath12k/dp_htt.h
@@ -225,7 +225,7 @@ struct htt_dp_umac_reset_setup_req_cmd {
  *          b'10:31 - reserved.
  */
 struct h2t_umac_hang_recovery_start_pre_reset {
-        u8 hdr;
+       u32 hdr;
 } __packed;
 
 #define HTT_H2T_UMAC_RESET_MSG_TYPE     GENMASK(7, 0)
